Added an appearing ul list with li-s
First added just a static ul list that had inside JS with .map method
that was dynamically adding li-s, but wanted to have a border that ape-
ared after first item was added and then grew. When added css to the
static ul, the border was visible as a line b4 adding the first li,
so tried to change that the ul was added too, not static, and did it
by adding another state variable, with an empty array, then after
running the add function, added one item, empty string, but not with the
[...item, ] thing, as wanted the array to have only one item in it, not
having a growing array. Casue wanted only one ul, and multiple li-s.