Fooked by vid probs, chest infec or something
was horizontal for 2 days straight, slept 20h a day, chest was mad,
but couldn't stop thinking wether I can rid off eventListeners, as it's
react with state. I did, it works. Brain fuzzy, so will take time, but
it works. Just need to change so after clikcin add butt it changes to
enter something bronx demon, and not enter job, to see if can do it.